requirements[1].txt
RPi.GPIO
Adafruit_DHT
telepot
